api/requirements.txt
fastapi
uvicorn
torch==2.6.0
#git+https://github.com/huggingface/transformers.git #transformers[torch]
transformers>=4.51.0
accelerate
spacy>=3.8
matplotlib
plotly==6.0.1
pandas
pyarrow
scikit-learn
typing-inspect
typing_extensions
pyyaml
types-PyYAML
protobuf
python-jose[cryptography]
bcrypt
requests
argparse
datasets
sentence-transformers<=5.1.0
sentencepiece
umap-learn
sacremoses
python-multipart
openai
sqlalchemy
python-slugify
psutil
openpyxl
fasttext @ git+https://github.com/facebookresearch/fastText.git@main
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0.tar.gz#egg=fr_core_news_sm
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz#egg=en_core_web_sm
ruff
black
mypy
types-python-jose
types-psutil
pandas-stubs
loky
psycopg2
python-dotenv
aiofiles
bertopic
stopwordsiso
hdbscan
einops
simplemma
nbformat
great_tables
pytz
faiss-cpu